requirements.txt
requests
python-ldap
pygerduty
